+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.9] - 2026-02-16
+
+### Fixed
+- Various bug fixes
+
+
+
 ## [0.4.8] - 2026-02-16
 
 ### Added
@@ -153,7 +160,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Release workflow accepts leading `v` in version input
 
 
-[Unreleased]: https://github.com/aarondfrancis/counselors/compare/v0.4.8...HEAD
+[Unreleased]: https://github.com/aarondfrancis/counselors/compare/v0.4.9...HEAD
 [0.3.0]: https://github.com/aarondfrancis/counselors/releases/tag/v0.3.0
 [0.3.1]: https://github.com/aarondfrancis/counselors/releases/tag/v0.3.1
 [0.3.2]: https://github.com/aarondfrancis/counselors/releases/tag/v0.3.2
@@ -168,3 +175,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.6]: https://github.com/aarondfrancis/counselors/releases/tag/v0.4.6
 [0.4.7]: https://github.com/aarondfrancis/counselors/releases/tag/v0.4.7
 [0.4.8]: https://github.com/aarondfrancis/counselors/releases/tag/v0.4.8
+[0.4.9]: https://github.com/aarondfrancis/counselors/releases/tag/v0.4.9